docker/pullrequestwatcher/requirements.txt
PyGithub==1.54.1
slackclient==2.9.3
